Export routes actions row to own component.
